feat: Add model for totoDev
This commit adds a new model called `totoDev` to the codebase. The `totoDev` model is defined with the necessary attributes such as `phone`, `role`, and `active`. This model is used to represent developers in the system and their roles. The addition of this model improves the functionality of the bot by allowing better management and tracking of developers.