feat: Add model for totoPlugin
This commit adds a new model called `totoPlugin` to the codebase. The `totoPlugin` model is defined with the necessary attributes such as `name`, `description`, `category`, and `usage`. This model is used to represent plugins in the system and their details. The addition of this model improves the functionality of the bot by allowing better management and tracking of plugins.